refactor: City component 로직 분리 완료
