+++ src/types/schedule.ts
@@ -1,6 +1,10 @@
-import { IDateProps } from './calendar';
+import { IDateProps, IOpenStateProps } from './calendar';
 import { ICityProps } from './city';
 
 export interface ICreateScheduleProps extends IDateProps {
+    city: ICityProps;
+}
+
+export interface IMakeTripDatesProps extends IDateProps, IOpenStateProps {
     city: ICityProps;
 }
